Before performing an upgrade, remove any previous copy of Growl and, if one was installed, ask the System Preferences to quit before proceeding.
On upgrade from 0.7.x previously:
(1) no pref pane update (2) error displayed (3) version not updated if the pref pane does manage to load (4) a duplicative "do you want to install this pref pane? it's newer!" message

The behavior is all right and clean as a result of this commit; System Preferences loads, installs the bundle, prompts for user or global installation, and displays the pref pane.