Deleted `try` block around `import _growl`.
At the very top of the Growl module was:

{{{
try:
    import _growl
except:
    _growl = False
}}}

With this try, we get this ''at registration'' if `_growl` is not present:

{{{
>>> import Growl
>>> n = Growl.GrowlNotifier('MyApp', ['Notification'])
[imagine five hours of runtime that doesn't involve using Growl]
>>> n.register()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "Growl.py", line 199, in register
    self._notifyMethod.PostNotification(GROWL_APP_REGISTRATION, regInfo)
AttributeError: 'bool' object has no attribute 'PostNotification'
}}}

There are three problems:

 1. The exception is disguised: The `import` is what failed, but the exception comes from `GrowlNotifier.register`.
 2. The exception happens on ''every'' call to `register`, rather than only once at `import`. This ties into the previous problem: The one `import` is what failed, but it's every `register` that gives the exception.
 3. The exception is the wrong class (`AttributeError` rather than `ImportError`) and its explanation is irrelevant to the actual problem.

So now the code is:

{{{
	import _growl
}}}

Do or do not. There is no try.

Here's what it says with this fix:

{{{
>>> import Growl
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "Growl.py", line 13, in <module>
    import _growl
ImportError: No module named _growl
}}}

This is much clearer. It comes right out and says the problem: ?No module named _growl?. This is the lead that a programmer trying to use the Python module would need to find out that he needs to install it first.