Fix validation. Really fixes #722.
First, I made GRDENotification use the notificationNames accessor I introduced in [4600] (well, actually [4599], but close enough). It's kinda dumb to be looking in an array of GRDENotifications for an NSString.

Second, in order for the validation method to actually get called, I needed to check a box (Validates Immediately) in IB (Tiger version, in case you're curious). Without this checked, KVB apparently does not call the bound object's KVV methods, immediately or otherwise (weirdly enough).