Added menu items to generate code to post the selected notifications. Fixes #622.
See GRDEDocument for a comment documenting this feature's implementation. One thing not mentioned there is that, currently, we only look in the app bundle for .grdecodesample files. In some future version, we could enable users to add .grdecodesample files in ~/Library/Application Support, but for now, they'll just have to put them in the app bundle.

Also, this change introduces an app delegate, with one outlet, named menuItemToInsertWriteCodeSampleToPasteboardMenuItemsAfter. (I mention this for searchability.)