GrowlPathUtilities' search for the GrowlHelperApp bundle now looks for a running GHA process first. If found, it returns the bundle for that process.
This fixes, framework-side, the bug where a bunch of .growlRegDict files get left in the Temporary Items folder. This was caused by GAB launching the .growlRegDict files with the GHA in ~/Library, when the GHA in /Library was the one running, and by GHA-in-~/Library bailing before it processed the files. Frameworks with this fix will use the running GHA, thereby correctly registering the app and not leaving behind a .growlRegDict in Temporary Items.