Because we can't create a PNG from a 1-bit image, we must ignore any 1-bit image reps.
In order to not ignore images that are *only* available in 1-bit, we now support TIFF as a fallback to PNG. When this happens, the data: URL changes to carry image/tiff instead of image/png as the media type, and to carry the TIFFRepresentation rather than the (nil) PNGRepresentation as the data.

Accordingly, I renamed -[NSImage GrowlBitmapImageRep] to -GrowlBitmapImageRepForPNG. It returns nil whenever the only available bitmap representations can't be converted to PNG (because they're 1-bit).

The previous code in -GrowlBitmapImageRep would have attempted to create a TIFFRepresentation from the 1-bit image rep, create an NSBitmapImageRep from the TIFFRepresentation, then create PNG data from the NSBitmapImageRep. I don't think this would have worked: it would only have worked if Cocoa would have forged RGB or 8-bit grayscale data at some point in the process, and I don't think it would have.