Fix #112: Speech rate not matching what's set in the Speech prefpane.
The fix is to create the speech synthesizer object using nil as the voice name, rather than asking for the default voice name. According to the NSSpeechSynthesizer documentation, the synthesizer object's initial rate is the rate set in Speech when the voice name is nil, whereas it's a blanket default when the voice name is the name of a specific voice.